Check for the sql-generator right away
