+++ cargo-pgx/src/command/pgx.rs
@@ -1,4 +1,6 @@
 use crate::CommandExecute;
+use std::path::Path;
+use owo_colors::OwoColorize;
 
 #[derive(clap::Args, Debug)]
 #[clap(about, author)]
@@ -34,6 +36,7 @@ enum CargoPgxSubCommands {
 impl CommandExecute for CargoPgxSubCommands {
     fn execute(self) -> eyre::Result<()> {
         use CargoPgxSubCommands::*;
+        check_for_sql_generator_binary()?;
         match self {
             Init(c) => c.execute(),
             Start(c) => c.execute(),
@@ -48,5 +51,22 @@ impl CommandExecute for CargoPgxSubCommands {
             Test(c) => c.execute(),
             Get(c) => c.execute(),
         }
+    }
+}
+
+/// A temporary check to help users from 0.2 or 0.3 know to take manual migration steps.
+fn check_for_sql_generator_binary() -> eyre::Result<()> {
+    if Path::new("src/bin/sql-generator.rs").exists() {
+        // We explicitly do not want to return a spantraced error here.
+        println!("{}", "\
+            Found `pgx` 0.2-0.3 series SQL generation while using `cargo-pgx` 0.4 series.
+            
+We've updated our SQL generation method, it's much faster! Please follow the upgrading steps listed in https://github.com/zombodb/pgx/releases/tag/v0.4.0.
+
+Already done that? You didn't delete `src/bin/sql-generator.rs` yet, so you're still seeing this message.\
+        ".red().bold());
+        std::process::exit(1)
+    } else {
+        Ok(())
     }
 }
